+++ src/strain_relief/data_types/__init__.py
@@ -1,6 +1,4 @@
 from ._types import (
-    ConfEnergiesDict,
-    ConformerASEList,
     EnergiesDict,
     MolPropertiesDict,
     MolsDict,
@@ -10,6 +8,4 @@ __all__ = [
     "MolsDict",
     "MolPropertiesDict",
     "EnergiesDict",
-    "ConfEnergiesDict",
-    "ConformerASEList",
 ]
