+++ tests/conftest.py
@@ -4,8 +4,8 @@ import numpy as np
 import pytest
 from strain_relief import test_dir
 from strain_relief.constants import EV_TO_KCAL_PER_MOL, MOL_KEY
+from strain_relief.data_types import EnergiesDict, MolPropertiesDict, MolsDict
 from strain_relief.io import load_parquet, to_mols_dict
-from strain_relief.types import EnergiesDict, MolPropertiesDict, MolsDict
 
 
 @pytest.fixture(scope="function")
